backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib
bcrypt
python-multipart
alembic
python-dotenv
pandas
openpyxl
xlrd
numpy